strings

+++ 0x00-python_variable_annotations/3-to_str.py
@@ -3,6 +3,6 @@
 """convert float to string"""
 
 
-def to_str(n: float) -> float:
+def to_str(n: float) -> int:
     """Takes a float n as argument and return the string of the float"""
     return str(float(n))
